Move AllConfusables component into its own directory
